% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predefined_model.R
\name{compare_viz_model}
\alias{compare_viz_model}
\title{Compare Model Performance with Visualization}
\usage{
compare_viz_model(
  x,
  stats = c("coef", "heatmap", "mosaic", "roc", "density", "bin"),
  nbins = 10,
  return = FALSE
)
}
\arguments{
\item{x}{list. multi_classifiers() 결과로 생성한 모델들의 list 객체.}

\item{stats}{character. 시각화할 플롯의 종류. 
c("coef", "heatmap", "mosaic", "roc", "density", "bin")에서 선택.}

\item{nbins}{integer. 예측 확률을 비닝할 때 사용할 빈(Bins)의 개수. 
stats의 값이 "quantiles"일 경우만 적용되며 기본값은 10.}

\item{return}{logical. 시각화 결과를 반환할지의 여부. 
기본값은 FALSE로 개별 결과를 list의 성분으로 반환하지 않음. 
TRUE이면 시각화 결과를 list로 반환.}
}
\value{
list. gg 클래스를 성분으로 갖는 리스트 객체.
}
\description{
multi_classifiers()로 생성한 여러 모델링의 성능 평가 및 비교를 위한 플롯을 시각화함
}
\details{
stats의 값에 따라 다음과 같은 플롯을 시각화함.
\itemize{
\item "coef": terms별 계수와의 관계 시각화.
\item "heatmap": Confusion Matrix 결과를 heatmap 플롯으로 시각화.
\item "heatmap": Confusion Matrix 결과를 mosaic 플롯으로 시각화.
\item "roc": ROC 커브 시각화.
\item "density": 예측확률의 target 클래스별 밀도 시각화.
\item "bin": 예측 확률의 구간별 target 클래스별 분포 시각화.
}
}
\examples{
\donttest{
# 모델 개발을 위한 recipe 정의
data_recipe <- dlookr::heartfailure \%>\% 
  target_to(death_event) \%>\% 
  set_positive("Yes") \%>\% 
  split_dataset(prop = 0.7) \%>\% 
  extract_dataset() \%>\% 
  set_formula() \%>\% 
  step_my_center() \%>\% 
  step_my_dummy() \%>\% 
  step_my_rose()
  
# "logistic", "lasso", "elastic", "ridge", "ranger", "xgboost" classifiers를 순차 모델링
my_models <- multi_classifiers(data_recipe)

# terms별 계수와의 관계 시각화
compare_viz_model(my_models, "coef")

# Confusion Matrix 결과를 heatmap 플롯으로 시각화
compare_viz_model(my_models, "heatmap")

# Confusion Matrix 결과를 mosaic 플롯으로 시각화
compare_viz_model(my_models, "mosaic")

# ROC 커브 시각화
compare_viz_model(my_models, "roc")

# 예측확률의 target 클래스별 밀도 시각화
compare_viz_model(my_models, "density")

# 예측 확률의 구간별 target 클래스별 분포 시각화
compare_viz_model(my_models, "bin")

# 예측 확률의 구간별 target 클래스별 분포 시각화 (15개 bins)
compare_viz_model(my_models, "bin", nbins = 15)

# terms별 계수와의 관계 시각화 결과 반환
compare_viz_model(my_models, "coef", return = TRUE)
}
}
\seealso{
\code{\link{compare_info_model}}
}
