% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda_category}
\alias{eda_category}
\title{범주형 변수의 excel EDA 보고서 생성}
\usage{
eda_category(
  x,
  target_variable,
  positive = "1",
  output_file = NULL,
  output_dir = getwd(),
  sample_percent = 100,
  parallel = FALSE,
  cores = parallel::detectCores() - 2,
  future_globals_maxsize = 500 * 1024^2,
  verbose = TRUE
)
}
\arguments{
\item{x}{data.frame, tbl_df. EDA를 수행할 데이터셋.}

\item{target_variable}{character. Target 변수의 이름.}

\item{positive}{character. Target 변수에서의 Positive 클래스의 이름}

\item{output_file}{character. 생성할 Excel 파일의 이름.}

\item{output_dir}{character. 생성할 Excel 파일이 저장될 디렉토리.}

\item{sample_percent}{numeric. 샘플로 분석할 경우의 샘플의 비율로서의 백분율.}

\item{parallel}{logical. 병렬 프로세싱으로 수행할지의 여부, 기본값은 FALSE.}

\item{cores}{integer. 병렬 프로세싱에서 사용할 코어의 개수.}

\item{future_globals_maxsize}{numeric. 병렬 프로세싱에서 개별 코어에서 
사용할 글로벌 메모리의 최대값.}

\item{verbose}{logical. 작업 경과의 정보를 출력할지의 여부, 기본값은 TRUE.}
}
\description{
범주형 변수에 대한 EDA 수행 결과를 Excel 파일로 생성함
}
\details{
범주형 변수의 총괄 정보를 제외한 모든 범주형 변수는 1개의 워크시트에 EDA 정보가 출력됨.
Excel은 256개의 워크시트만 지원하므로 범주형 변수 개수가 255개를 초과하는 것은 255개의 정보만 생성됨.
}
\examples{
\dontrun{
testdata <- dlookr::heartfailure
target_variable <- "death_event"

# single core processing
eda_category(testdata, target_variable, positive = "Yes")

# 30\% 샘플링한 데이터로 보고서를 생성함
eda_category(testdata, target_variable, positive = "Yes", sample_percent = 30)

# parallel processing
eda_category(testdata, target_variable, positive = "Yes", parallel = TRUE,
             cores = 8)
}

}
\seealso{
\code{\link{eda_category}}
}
\author{
유충현
Maintainer: 유충현 <antony@hanwha.com>
}
