% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_by.R
\name{extract_dataset}
\alias{extract_dataset}
\title{데이터셋 추출}
\usage{
extract_dataset(.data, set = "train")
}
\arguments{
\item{.data}{a \code{\link{initial_split}}.}

\item{set}{character. 데이터셋의 종류. "train" 혹은 "test"로 설정. 
기본값은 "train"으로 모델 개발을 위한 training 셋을 추출함.}
}
\value{
grouped_df 클래스 객체.
grouped_df 클래스 객체의 기본 속성 외의, 속성은 다음과 같음.
\itemize{
\item target character. Target 변수 이름.
\item positive character. positive 클래스 이름.
}
}
\description{
initial_split 클래스 객체로부터 모델 개발을 위한 training 셋이나 모델 평가를 위한 test 셋을 추출
}
\examples{
\donttest{
library(dplyr)

# training 데이터셋을 70\%로 데이터 분리 설정
data_targeted <- dlookr::heartfailure \%>\% 
  target_to(death_event) \%>\% 
  set_positive("Yes") \%>\% 
  split_dataset(prop = 0.7) \%>\% 
  extract_dataset()
  
data_targeted   
}
}
