% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predefined_model.R
\name{get_classifier}
\alias{get_classifier}
\title{모델 메타 정의}
\usage{
get_classifier(
  model = c("logistic", "lasso", "elastic", "ridge", "ranger", "xgboost")
)
}
\arguments{
\item{model}{character. 모델의 이름. c("logistic", "lasso", "elastic", "ridge", 
"ranger", "xgboost")에서 선택.}
}
\value{
model_spec 클래스 객체.
}
\description{
parsnip 패키지를 이용해서, tidymodels 에코 시스템에서 사용할 수 있는 모델 정의
}
\examples{
\donttest{
get_classifier("logistic")

get_classifier("lasso")

get_classifier("ridge") 

get_classifier("elastic") 

get_classifier("ranger") 

get_classifier("xgboost") 
}

}
