% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predefined_model.R
\name{multi_classifiers}
\alias{multi_classifiers}
\title{Fit Binary Classification with Multi Classifiers}
\usage{
multi_classifiers(
  x,
  classifiers = c("logistic", "lasso", "elastic", "ridge", "ranger", "xgboost"),
  n_fold = 10,
  n_grid = 30,
  n_best = 10,
  n_trees = ifelse(classifiers \%in\% "ranger", 500L, 500L),
  best_metric = c("f_meas", "recall", "sensitivity", "precision", "specificity",
    "accuracy", "bal_accuracy", "detection_prevalence", "f_meas_05", "f_meas_2",
    "j_index", "kap", "mcc", "npv", "ppv", "gmean"),
  parallel = FALSE,
  cores = 6,
  future_globals_maxsize = 500 * 1024^2,
  verbose = TRUE
)
}
\arguments{
\item{x}{recipe. 모델을 정의한 recipe 객체.}

\item{classifiers}{character. 모델의 classifiers 이름을 담은 문자열 벡터. 
c("logistic", "lasso", "elastic", "ridge", "ranger", "xgboost")에서 복수 선택.}

\item{n_fold}{integer. Cross-Validation을 수행한 n-Folds의 값. 기본값은 10.}

\item{n_grid}{integer. Hyper Parameters를 수행할 Grid의 개수. 기본값은 30.}

\item{n_best}{integer. Best 모델의 목록을 계산할 목록의 개수. 기본값은 10.}

\item{n_trees}{integer. model이 "ranger"나 "xgboost"일 경우에만 사용하는 인수.
모델에서 생성할 개별 트리의 개수를 지정함. 기본값은 "ranger"는 500개, "xgboost"는 500개임.}

\item{best_metric}{character. Best 모델을 선정할 때 사용하는 성능지표. 
c("f_meas", "recall", "sensitivity", "precision", "specificity", "accuracy", 
"bal_accuracy", "detection_prevalence", "f_meas_05", "f_meas_2", "j_index", 
"kap", "mcc", "npv", "ppv", "gmean") 중에서 선택함. 기본값은 "f_meas".}

\item{parallel}{logical. 병렬 프로세싱으로 수행할지의 여부, 기본값은 FALSE.}

\item{cores}{integer. 병렬 프로세싱에서 사용할 코어의 개수.}

\item{future_globals_maxsize}{numeric. 병렬 프로세싱에서 개별 코어에서 
사용할 글로벌 메모리의 최대값.}

\item{verbose}{logical. 작업 경과의 정보를 출력할지의 여부, 기본값은 TRUE.}
}
\value{
list. workflow 클래스 객체를 성분으로 갖는 list.
개별 성분으로 반환하는 workflow 객체는 다음과 같은 attributes를 갖음.
\itemize{
\item coef_term tbl_df. 모델을 정의한 terms의 추정 계수값 등의 정보. 
\item pred tbl_df. 모델에 test 셋을 적용해서 예측한 예측값 정보. 
\item cmat conf_mat. 모델의 성능 측정을 위한 confusion matrix.
\item metrics tbl_df. 모델의 성능 측정 메트릭 정보.
}
}
\description{
여러 개의 classifiers로 Binary Classification modeling을 수행함
}
\examples{
\donttest{
# 모델 개발을 위한 recipe 정의
data_recipe <- dlookr::heartfailure \%>\% 
  target_to(death_event) \%>\% 
  set_positive("Yes") \%>\% 
  split_dataset(prop = 0.7) \%>\% 
  extract_dataset() \%>\% 
  set_formula() \%>\% 
  step_my_center() \%>\% 
  step_my_dummy() \%>\% 
  step_my_rose()
  
# "logistic", "lasso", "elastic", "ridge", "ranger", "xgboost" classifiers를 순차 모델링
my_models <- multi_classifiers(data_recipe)

# "lasso", "elastic", "ridge" classifiers를 순차 모델링
my_models <- multi_classifiers(data_recipe, 
                               classifiers = c("lasso", "elastic", "ridge"))

# "logistic", "lasso", "elastic", "ridge", "ranger", "xgboost"  classifiers를 병렬 모델링
my_models <- multi_classifiers(data_recipe, parallel = TRUE)
}
}
