% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{report_model}
\alias{report_model}
\title{적합된 모델 성능 보고서 생성}
\usage{
report_model(x, output_file = NULL, output_dir = getwd())
}
\arguments{
\item{x}{workflow, 혹은 list. 성능 보고서를 생성할 모델 적합 객체. 
단일 classifier일 경우에는 workflow 객체이고, 다중 classifiers일 경우에는 list 객체임.}

\item{output_file}{character. 생성할 Excel 파일의 이름.}

\item{output_dir}{character. 생성할 Excel 파일이 저장될 디렉토리.}
}
\description{
적합한 모델의 classifiers별로 성능 평가 결과를 Excel 파일로 생성함
}
\details{
적합된 모델들의 성능을 비교하기 위한 여러 통계 정보화 플롯을 Excel 파일에 출력함.
classifiers별로 17개의 성능 평가지표를 비교할 수 있는 총괄 워크시트와 
개별 classifiers의 상세 정보를 워크시트로 갖는 파일을 생성함.
}
\examples{
\dontrun{
# 모델 개발을 위한 recipe 정의
data_recipe <- dlookr::heartfailure \%>\% 
  target_to(death_event) \%>\% 
  set_positive("Yes") \%>\% 
  split_dataset(prop = 0.7) \%>\% 
  extract_dataset() \%>\% 
  set_formula() \%>\% 
  step_my_center() \%>\% 
  step_my_dummy() \%>\% 
  step_my_rose()

# logistic regrssion 적합     
my_model <- get_training(data_recipe)

# logistic regrssion 성능평가 보고서 출력
report_model(my_model)
   
# 복수개의 classifiers별로 적합
# "logistic", "lasso", "elastic", "ridge", "ranger", "xgboost" classifiers를 순차 모델링
my_models <- multi_classifiers(data_recipe)

# classifiers별로 성능을 비교할 수 있는 성능평가 보고서 출력
report_model(my_models, output_file = "compare_model_performance.xlsx")
}

}
\seealso{
\code{\link{info_model}}, \code{\link{viz_model}}, \code{\link{compare_info_model}}, \code{\link{compare_viz_model}}
}
\author{
유충현
Maintainer: 유충현 <antony@hanwha.com>
}
