% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{report_model_recipes}
\alias{report_model_recipes}
\title{적합된 Multi Recipes 모델 성능 보고서 생성}
\usage{
report_model_recipes(
  x,
  output_file = NULL,
  output_dir = getwd(),
  verbose = TRUE
)
}
\arguments{
\item{x}{list. multi_recipes() 함수의 결과로, workflow 클래스 객체를 성분으로 
갖는 list를 성분으로 갖는 list 객체.}

\item{output_file}{character. 생성할 Excel 파일의 이름. 기본값은 "model_performance.xlsx".}

\item{output_dir}{character. 생성할 Excel 파일이 저장될 디렉토리. 기본값은 현재 작업 디렉토리.}

\item{verbose}{logical. 작업 경과의 정보를 출력할지의 여부, 기본값은 TRUE.}
}
\description{
적합한 Multi Recipes 모델의 개별 recipes에 대해서 classifiers별로 
성능 평가 결과를 Excel 파일로 생성함
}
\details{
output_file의 기본값 "model_performance.xlsx"일 경우에 파일의 이름은
"model_performance_target변수이름_순번.xlsx"으로 생성됨
}
\examples{
\dontrun{
# 모델 개발을 위한 첫째 recipe 정의
death_event_recipe <- dlookr::heartfailure \%>\%
  target_to(death_event) \%>\%
  set_positive("Yes") \%>\%
  split_dataset(prop = 0.7) \%>\%
  extract_dataset() \%>\%
  set_formula() \%>\%
  step_my_center() \%>\%
  step_my_dummy() \%>\%
  step_my_rose()
  
# 모델 개발을 위한 둘째 recipe 정의
diabetes_recipe <- dlookr::heartfailure \%>\%
  target_to(diabetes) \%>\%
  set_positive("Yes") \%>\%
  split_dataset(prop = 0.7) \%>\%
  extract_dataset() \%>\%
  set_formula() \%>\%
  step_my_center() \%>\%
  step_my_dummy() \%>\%
  step_my_rose()

# recipe를 성분으로 갖는 list 정의
recipes <- list(death_event_recipe, diabetes_recipe)

my_recipes <- multi_recipes(recipes)

# recipes별로 classifiers의 성능을 비교할 수 있는 성능평가 보고서 출력
report_model_recipes(my_recipes)

# 생성된 파일의 이름들
list.files(path = getwd(), pattern = "model_performance")

}
}
\seealso{
\code{\link{report_model}}, \code{\link{multi_recipes}}
}
\author{
유충현
Maintainer: 유충현 <antony@hanwha.com>
}
