% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_recipes.R
\name{set_formula}
\alias{set_formula}
\title{모델 포뮬러 설정}
\usage{
set_formula(.data, formula = NULL)
}
\arguments{
\item{.data}{a \code{\link{grouped_df}}.}

\item{formula}{모델을 정의하는 formula. 기본값은 `target 변수 ~ .`}
}
\value{
recipe 클래스 객체.
}
\description{
Target 변수를 예측하기위한 formula를 지정함
}
\examples{
\donttest{
library(dplyr)

# training 데이터셋을 70\%로 데이터 분리 설정
data_recipe <- dlookr::heartfailure \%>\% 
  target_to(death_event) \%>\% 
  set_positive("Yes") \%>\% 
  split_dataset(prop = 0.7) \%>\% 
  extract_dataset() \%>\% 
  set_formula()
  
data_recipe  
}

}
