% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_by.R
\name{set_positive}
\alias{set_positive}
\title{Positive 클래스 지정}
\usage{
set_positive(.data, positive)
}
\arguments{
\item{.data}{a model_df.}

\item{positive}{character. positive 클래스.}
}
\value{
model_df 클래스 객체.
model_df 클래스 객체의 속성은 다음과 같음.
\itemize{
\item target character. Target 변수 이름.
\item positive character. positive 클래스 이름.
}
}
\description{
Binary 클래스를 갖는 Target 변수에 Positive 클래스를 지정함
}
\examples{
\donttest{
library(dplyr)

data_targeted <- dlookr::heartfailure \%>\% 
  target_to(death_event) \%>\% 
  set_positive("Yes")
  
data_targeted   
}

}
