% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_by.R
\name{split_dataset}
\alias{split_dataset}
\title{데이터셋 분리를 위한 설정}
\usage{
split_dataset(.data, prop = 3/4, breaks = 4, pool = 0.1, seed = NULL)
}
\arguments{
\item{.data}{a model_df.}

\item{prop}{numeric. training 데이터셋의 비율.}

\item{breaks}{numeric. 수치형 계층화 변수를 계층화하는데 필요한 bins의 개수를 지정하는 단일 숫자.}

\item{pool}{numeric. 특정 그룹이 너무 작아서 다른 그룹으로 모아야 하는지 판단하는데
사용되는, 데이터의 비율. 너무 작은 그룹을 계층화할 위험이 있으므로 이 인수를 
기본값인 0.1 미만으로 낮추는 것을 권장하지 않음.}

\item{seed}{integer. 데이터셋을 분리할 때의 샘플링 난수 시드.}
}
\value{
rsample 패키지에서 사용하는 initial_split 클래스 객체.
}
\description{
데이터셋을 모델 개발을 위한 training 셋과 모델 평가를 위한 test 셋으로 분리하는 설정
}
\examples{
\donttest{
library(dplyr)

# training 데이터셋을 70\%로 데이터 분리 설정
data_targeted <- dlookr::heartfailure \%>\% 
  target_to(death_event) \%>\% 
  set_positive("Yes") \%>\% 
  split_dataset(prop = 0.7)
  
data_targeted   
}
}
