% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_recipes.R
\name{step_my_center}
\alias{step_my_center}
\title{수치 데이터의 정규화}
\usage{
step_my_center(
  .recipe,
  ...,
  role = NA,
  trained = FALSE,
  means = NULL,
  na_rm = TRUE,
  skip = FALSE,
  id = rand_id("center")
)
}
\arguments{
\item{role}{새 변수가 생성되지 않으므로, 이 단계에서는 사용하지 않음.}

\item{trained}{logical. 전처리로 수치를 추정했는지의 여부.}

\item{means}{numeric. 평균. 이 값은 prep()에 의해서 계산될 때까지 NULL임.}

\item{na_rm}{logical. 계산할 떄, NA를 제거해야 하는지의 여부.}

\item{skip}{logical. 레시피가 bake()에 의해 전처리가 수행될 때 단계를 건너뛰어야 합니까? 
prep()가 실행될 때, 모든 전처리 작업이 수행되는 동안 일부 작업은 새 데이터에서 수행되지 않을 수 있습니다(예: 결과 변수 처리). 
skip = TRUE은 후속 작업의 계산에 영향을 미칠 수 있으므로 사용 시 주의해야 함.}

\item{id}{character. 식별하기 위한 고유값.}

\item{.data}{recipe 클래스 객체. 이 레시피에 해당작업 시퀀스가 추가됨.}
}
\value{
recipe 클래스 객체.
}
\description{
평균이 0이 되도록 수치 변수를 정규화하는 레시피 스펙 정의
}
\examples{
\donttest{
library(dplyr)

# training 데이터셋을 70\%로 데이터 분리 설정
data_recipe <- dlookr::heartfailure \%>\% 
  target_to(death_event) \%>\% 
  set_positive("Yes") \%>\% 
  split_dataset(prop = 0.7) \%>\% 
  extract_dataset() \%>\% 
  set_formula() \%>\% 
  step_my_center() 
  
data_recipe 

data_recipe \%>\% 
  prep()
    
data_recipe \%>\% 
  prep() \%>\% 
  juice() 
}

}
