% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_recipes.R
\name{step_my_dummy}
\alias{step_my_dummy}
\title{범주형 데이터의 더미화}
\usage{
step_my_dummy(
  .recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  one_hot = FALSE,
  naming = dummy_names,
  levels = NULL,
  keep_original_cols = FALSE,
  skip = FALSE,
  id = rand_id("dummy")
)
}
\arguments{
\item{role}{기본적으로 이 단계로 생성된 새 변수는 예측변수로 사용됨.}

\item{trained}{logical. 전처리로 수치를 추정했는지의 여부.}

\item{one_hot}{logical. 원-핫 인코딩 여부.}

\item{naming}{function. 새 더미 변수에 대한 명명규칙을 정의하는 함수.}

\item{levels}{list. 각 변수에 대해서 더미변수를 생성하는데 필요한 정보가 포함된 list.}

\item{keep_original_cols}{logical. 출력 결과에 원래의 변수를 유지하는지의 여부. 기본값은 FALSE.}

\item{skip}{logical. 레시피가 bake()에 의해 전처리가 수행될 때 단계를 건너뛰어야 합니까? 
prep()가 실행될 때, 모든 전처리 작업이 수행되는 동안 일부 작업은 새 데이터에서 수행되지 않을 수 있습니다(예: 결과 변수 처리). 
skip = TRUE은 후속 작업의 계산에 영향을 미칠 수 있으므로 사용 시 주의해야 함.}

\item{id}{character. 식별하기 위한 고유값.}

\item{.data}{recipe 클래스 객체. 이 레시피에 해당작업 시퀀스가 추가됨.}
}
\value{
recipe 클래스 객체.
}
\description{
명목척도의 범주형 변수를 하나 이상의 이진항으로 변환하는 레시피 스펙 정의
}
\examples{
\donttest{
library(dplyr)

# training 데이터셋을 70\%로 데이터 분리 설정
data_recipe <- dlookr::heartfailure \%>\% 
  target_to(death_event) \%>\% 
  set_positive("Yes") \%>\% 
  split_dataset(prop = 0.7) \%>\% 
  extract_dataset() \%>\% 
  set_formula() \%>\% 
  step_my_center() \%>\% 
  step_my_dummy()
  
data_recipe 

data_recipe \%>\% 
  prep()
    
data_recipe \%>\% 
  prep() \%>\% 
  juice() 
}

}
