% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_recipes.R
\name{step_my_rose}
\alias{step_my_rose}
\title{ROSE 알고리즘 적용}
\usage{
step_my_rose(
  .recipe,
  ...,
  role = NA,
  trained = FALSE,
  column = NULL,
  over_ratio = 1,
  minority_prop = 0.5,
  minority_smoothness = 1,
  majority_smoothness = 1,
  skip = TRUE,
  seed = sample.int(10^5, 1),
  id = rand_id("rose")
)
}
\arguments{
\item{.recipe}{recipe 객체. 이 recipe 객체에 대한 ROSE 순차작업을 추가함.}

\item{trained}{logical. quantities에 대해서 전처리 추정 여부를 나타냄.}

\item{column}{character. ... 선택자에 의해서 채워질 변수의 이름.}

\item{over_ratio}{numeric. majority-to-minority 빈도의 비율에 대한 숫자 값.
기본값(1)은 다른 모든 수준이 가장 많이 발생하는 수준과 동일한 빈도를 갖도록 샘플링됨을 의미. 
0.5 값은 minority 수준이 majority 수준보다 (최대) (대략) 절반의 행을 갖는다는 것을 의미.}

\item{minority_prop}{numeric. minority 클래스의 오버 샘플링 여부를 결정. 기본값은 0.5.}

\item{minority_smoothness}{numeric. minority 클래스의 조건부 커널 밀도를 
추정하기 위해 스무딩 매개변수를 곱할 축소 계수. 기본값은 1.}

\item{majority_smoothness}{numeric. majority 클래스의 조건부 커널 밀도를 
추정하기 위해 스무딩 매개변수를 곱할 축소 계수. 기본값은 1.}

\item{skip}{logical. 레시피가 bake()에 의해 전처리가 수행될 때 단계를 건너뛰어야 합니까? 
prep()가 실행될 때, 모든 전처리 작업이 수행되는 동안 일부 작업은 새 데이터에서 수행되지 않을 수 있습니다(예: 결과 변수 처리). 
skip = TRUE은 후속 작업의 계산에 영향을 미칠 수 있으므로 사용 시 주의해야 함.}

\item{seed}{integer. 시드 값.}

\item{id}{character. 식별하기 위한 고유값.}

\item{role.}{새 변수가 생성되지 않으므로, 이 단계에서는 사용하지 않음.}
}
\value{
recipe 객체.
}
\description{
minority와 majority 클래스를 확대하여, 
ROSE(Randomly Over Sampling Examples) 기반 합성 데이터 샘플을 생성하는 
레시피 스펙을 정의함. ROSE::ROSE()를 사용함.
}
\details{
themis::step_rose() 도움말 참고.
}
\examples{
# Generate data for the example
\donttest{
data_recipe <- dlookr::heartfailure \%>\% 
  target_to(death_event) \%>\% 
  set_positive("Yes") \%>\% 
  split_dataset(prop = 0.7) \%>\% 
  extract_dataset() \%>\% 
  set_formula() \%>\% 
  step_my_center() \%>\% 
  step_my_dummy() \%>\% 
  step_my_rose()
  
data_recipe \%>\% 
  recipes::prep() \%>\% 
  recipes::juice()
}

}
\author{
유충현
Maintainer: 유충현 <antony@hanwha.com>
}
