% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_by.R
\name{target_table}
\alias{target_table}
\title{target 변수 분포 출력}
\usage{
target_table(.data)
}
\arguments{
\item{.data}{model_df, initial_split 혹은 recipe. 분포를 출력할 대상 객체.}
}
\value{
tbl_df.
\itemize{
  \item Target 변수
  \item 건수
  \item 백분율
}
}
\description{
데이터셋 중 target 변수의 Positive와 Negative의 분포 출력
}
\details{
tidymodels 패키지군 기반으로 모델링을 하는 단계에서, 
원 데이터와 전처리 후의 데이터의 Target 변수에서의 Positive와 Negative의 분포 출력
}
\examples{
\dontrun{
data_recipe <- dlookr::heartfailure \%>\% 
  target_to(death_event) 

# model_df 클래스에 대한 원 데이터의 분포
target_table(data_recipe)

data_recipe <- dlookr::heartfailure \%>\% 
  target_to(death_event) \%>\% 
  set_positive("Yes") \%>\% 
  split_dataset(prop = 0.7)

# initial_split 클래스에 대한, 분할하였을 때 Traning 셋의 데이터의 분포
target_table(data_recipe)

data_recipe <- dlookr::heartfailure \%>\% 
  target_to(death_event) \%>\% 
  set_positive("Yes") \%>\% 
  split_dataset(prop = 0.7) \%>\% 
  extract_dataset()

# initial_split 클래스에, 대한 분할 후의 Traning 셋의 데이터의 분포
target_table(data_recipe)

data_recipe <- dlookr::heartfailure \%>\% 
  target_to(death_event) \%>\% 
  set_positive("Yes") \%>\% 
  split_dataset(prop = 0.7) \%>\% 
  extract_dataset() \%>\% 
  set_formula() \%>\% 
  step_my_rose()

# recipe 클래스에 대한 전처리 후의 데이터의 분포
target_table(data_recipe)
}

}
\author{
유충현
Maintainer: 유충현 <antony@hanwha.com>
}
