% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_by.R
\name{target_to}
\alias{target_to}
\alias{target_to.data.frame}
\title{Target 변수 지정}
\usage{
target_to(.data, target, ...)

\method{target_to}{data.frame}(.data, target, ...)
}
\arguments{
\item{.data}{a data.frame or a \code{\link{tbl_df}}.}

\item{target}{target 변수 이름.}

\item{...}{arguments to be passed to methods.}
}
\value{
model_df 클래스 객체.
model_df 클래스 객체의 속성은 다음과 같음.
\itemize{
\item target character. Target 변수 이름.
}
}
\description{
Machine learning 과정에서 Target 변수를 지정함
}
\examples{
\donttest{
library(dplyr)

data_targeted <- dlookr::heartfailure \%>\% 
  target_to(death_event)
  
data_targeted   
}

}
