% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predefined_model.R
\name{viz_model}
\alias{viz_model}
\title{Visualization for Model Performance}
\usage{
viz_model(
  x,
  stats = c("coef", "heatmap", "mosaic", "roc", "density", "bin"),
  nbins = 10
)
}
\arguments{
\item{x}{workflow. get_training() 결과로 생성한 모델, workflow 객체.}

\item{stats}{character. 시각화할 플롯의 종류. 
c("coef", "heatmap", "mosaic", "roc", "density", "bin")에서 선택.}

\item{nbins}{integer. 예측 확률을 비닝할 때 사용할 빈(Bins)의 개수. 
stats의 값이 "quantiles"일 경우만 적용되며 기본값은 10.}
}
\value{
gg 클래스의 객체. ggplot2 시각화 객체
}
\description{
모델링 함수로 생성한 모델에 대해서 성능 평가를 위한 플롯을 시각화함
}
\details{
stats의 값에 따라 다음과 같은 플롯을 시각화함.
\itemize{
\item "coef": terms별 계수와의 관계 시각화. tree 기반의 모델링에서는 계수가 아닌, feature importance 값임.
\item "heatmap": Confusion Matrix 결과를 heatmap 플롯으로 시각화.
\item "heatmap": Confusion Matrix 결과를 mosaic 플롯으로 시각화.
\item "roc": ROC 커브 시각화.
\item "density": 예측확률의 target 클래스별 밀도 시각화.
\item "bin": 예측 확률의 구간별 target 클래스별 분포 시각화.
}
}
\examples{
\donttest{
# 모델 개발을 위한 recipe 정의
data_recipe <- dlookr::heartfailure \%>\% 
  target_to(death_event) \%>\% 
  set_positive("Yes") \%>\% 
  split_dataset(prop = 0.7) \%>\% 
  extract_dataset() \%>\% 
  set_formula() \%>\% 
  step_my_center() \%>\% 
  step_my_dummy() \%>\% 
  step_my_rose()

# 정의한 recipe로 모델 빌드
my_model <- get_training(data_recipe)

# terms별 계수와의 관계 시각화
viz_model(my_model, "coef")

# Confusion Matrix 결과를 heatmap 플롯으로 시각화
viz_model(my_model, "heatmap")

# Confusion Matrix 결과를 mosaic 플롯으로 시각화
viz_model(my_model, "mosaic")

# ROC 커브 시각화
viz_model(my_model, "roc")

# 예측확률의 target 클래스별 밀도 시각화
viz_model(my_model, "density")

# 예측 확률의 구간별 target 클래스별 분포 시각화
viz_model(my_model, "bin")

# 예측 확률의 구간별 target 클래스별 분포 시각화 (15개 bins)
viz_model(my_model, "bin", nbins = 15)
}

}
\seealso{
\code{\link{info_model}}
}
